requirements.txt
psycopg2-binary==2.9.10
pymongo==4.11.1
mysql-connector==2.2.9
pymysql>=1.1.1
sqlalchemy>=2.0.39